Allow multiple possible library paths
Currently, if the cumulative total length of all possible library
paths is greater than the maximum length of a single path (on this
OS), then the path to the library is left blank. This results in
failure to load the library.

This fix is the all-platforms port for the PR linked below, and
enables each separate path to be measured for length independent of
its neighbours.

https://github.com/ibmruntimes/openj9-openjdk-jdk8/pull/107

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>